1) A new module for dependency injection has been added - ui module.    CompositionsLocal now only provides di modules for easy access. 2) Fixed milliliters resolving bug for water tracks in water tracking feature. 3) Two new fields have been added to the water tracking statistics - the average percentage of water in drinks and the average number of water drinks    per day.    Incorrect favorite drink in statistics has also been fixed. 4) A new Composable section (PermissionsSection) has been added to uikit for features that require permissions. 5) The pedometer manager has been slightly redesigned. 6) New string resources and several json lottie animations have been added and localized. 7) The project structure has been slightly refactored.

+++ app/logic/src/main/kotlin/hardcoder/dev/logic/features/waterBalance/WaterPercentageResolver.kt
@@ -5,6 +5,6 @@ import hardcoder.dev.entities.features.waterTracking.DrinkType
 class WaterPercentageResolver {
 
     fun resolve(drinkType: DrinkType, millilitersDrunk: Int): Int {
-        return  millilitersDrunk / 100 * drinkType.hydrationIndexPercentage
+        return (millilitersDrunk.toFloat() / 100F * drinkType.hydrationIndexPercentage.toFloat()).toInt()
     }
 }
